🧹 Refactor send_notifications into distinct handlers
- Extracted monolithic `_send` logic in `backend/routers/events.py` to a new `notification_service.py` module to match the existing codebase service pattern.
- Separated logic into distinct functional handlers (`_send_telegram`, `_send_email`, `_send_webhook`) for improved maintainability.
- Maintained background threading execution and all existing functionalities/configurations.
- Verified changes with ruff linter and py_compile.
